Refactor Puppet Commands in extension.ts
Move out all of the puppet commands to seperate files and make
extension.ts simplier.

This removes alot of the imports in extension.ts and cleans up the
method explosion.
