+++ client/src/extension.ts
@@ -1,25 +1,11 @@
 'use strict';
 
 import * as vscode from 'vscode';
-import * as path from 'path';
-import * as net from 'net';
-import * as cp from 'child_process';
-import ChildProcess = cp.ChildProcess;
-import {
-  LanguageClient, LanguageClientOptions, SettingMonitor, ServerOptions,
-  ErrorAction, ErrorHandler, CloseAction, TransportKind, RequestType0
-} from 'vscode-languageclient';
 
-import {
-  PuppetNodeGraphContentProvider, isNodeGraphFile, getNodeGraphUri,
-  showNodeGraph, getViewColumn
-} from '../src/providers/previewNodeGraphProvider';
-import { puppetResourceCommand } from '../src/commands/puppetResourceCommand';
-import { puppetModuleCommand } from '../src/commands/puppetModuleCommand';
-import * as messages from '../src/messages';
 import { startLangServerTCP } from '../src/languageserver';
+import { setupPuppetCommands } from '../src/puppetcommands';
 
-const langID = 'puppet';
+const langID = 'puppet'; // don't change this
 var statusBarItem;
 
 export function activate(context: vscode.ExtensionContext) {
@@ -33,35 +19,10 @@ export function activate(context: vscode.ExtensionContext) {
 
   createStatusBarItem();
 
-  var languageServerClient = startLangServerTCP(host, port, langID, [langID], statusBarItem);
+  var languageServerClient = startLangServerTCP(host, port, langID, statusBarItem);
   context.subscriptions.push(languageServerClient.start());
 
-  let resourceCommand = new puppetResourceCommand(languageServerClient);
-  context.subscriptions.push(resourceCommand);
-  context.subscriptions.push(vscode.commands.registerCommand('extension.puppetResource', () => {
-    resourceCommand.run();
-  }));
-
-  let moduleCommand = new puppetModuleCommand();
-  context.subscriptions.push(moduleCommand);
-  context.subscriptions.push(vscode.commands.registerCommand('extension.puppetModule', () => {
-    moduleCommand.listModules();
-  }));
-
-  context.subscriptions.push(vscode.commands.registerCommand(
-    'extension.puppetShowNodeGraphToSide',
-    uri => showNodeGraph(uri, true))
-  );
-
-  const contentProvider = new PuppetNodeGraphContentProvider(context, languageServerClient);
-  const contentProviderRegistration = vscode.workspace.registerTextDocumentContentProvider(langID, contentProvider);
-
-  context.subscriptions.push(vscode.workspace.onDidSaveTextDocument(document => {
-    if (isNodeGraphFile(document)) {
-      const uri = getNodeGraphUri(document.uri);
-      contentProvider.update(uri);
-    }
-  }));
+  setupPuppetCommands(langID, languageServerClient, context);
 
   console.log('Congratulations, your extension "vscode-puppet" is now active!');
 }
